% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_conda_availability_from_yaml}
\alias{check_conda_availability_from_yaml}
\title{Check which R packages in environment.yml are not available on conda-forge}
\usage{
check_conda_availability_from_yaml(file = "environment.yml")
}
\arguments{
\item{file}{Path to environment.yml (default: "environment.yml")}
}
\value{
A character vector of missing packages
}
\description{
Check which R packages in environment.yml are not available on conda-forge
}
